pyproject.toml
[project]
name = "urp-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"flask>=3.1.3",
	"requests>=2.32.5",
	"ddddocr>=1.6.1",
	"pillow>=12.1.1",
	"beautifulsoup4>=4.14.2",
	"pycryptodome>=3.23.0",
]


